btrfs-clone: support subvolumes with spaces
Add support for subvolumes which might have spaces in their name. The
subvolume name is the last field in the subvolume list so if whitespace
is present after the start of this field it is part of the subvolume
name. This sets maxsplit=3 so that it will be split into 4 fields.